# server/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentConfirmation": {
        "additionalProperties": false,
        "properties": {
          "approve": {
            "title": "Approve",
            "type": "boolean"
          },
          "idempotency_key": {
            "maxLength": 192,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "payload_digest": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Payload Digest",
            "type": "string"
          },
          "work_item_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Work Item Id",
            "type": "string"
          }
        },
        "required": [
          "work_item_id",
          "approve",
          "payload_digest",
          "idempotency_key"
        ],
        "title": "AgentConfirmation",
        "type": "object"
      },
...
      "AgentInteractionAnswerRequest": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Answers",
            "type": "array"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 192,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "required": [
          "answers"
        ],
        "title": "AgentInteractionAnswerRequest",
        "type": "object"
      },
...
      "ConfirmWorkResponse": {
        "properties": {
          "payloadDigest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payloaddigest"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "workItemId": {
            "title": "Workitemid",
            "type": "string"
          }
        },
        "required": [
          "status",
          "workItemId"
        ],
        "title": "ConfirmWorkResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InteractionAnswerResponse": {
        "properties": {
          "interactionId": {
            "title": "Interactionid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "interactionId"
        ],
        "title": "InteractionAnswerResponse",
        "type": "object"
      },
...
      "QuizSubmissionBody": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 192,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "submission_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "submission_id",
          "answers"
        ],
        "title": "QuizSubmissionBody",
        "type": "object"
      },
      "QuizSubmissionResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuizSubmissionSnapshotInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "status"
        ],
        "title": "QuizSubmissionResponse",
        "type": "object"
      },
      "QuizSubmissionSnapshotInfo": {
        "properties": {
          "handoff_reason": {
            "default": "",
            "title": "Handoff Reason",
            "type": "string"
          },
          "per_question": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Per Question",
            "type": "array"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "total_points": {
            "default": 0,
            "title": "Total Points",
            "type": "integer"
          },
          "total_score": {
            "default": 0.0,
            "title": "Total Score",
            "type": "number"
          }
        },
        "required": [
          "submission_id"
        ],
        "title": "QuizSubmissionSnapshotInfo",
        "type": "object"
      },
...
      "SchedulePermissionDecision": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "allow",
              "allow_chat",
              "always_allow",
              "skip"
            ],
            "title": "Decision",
            "type": "string"
          },
          "proposalId": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Proposalid",
            "type": "string"
          }
        },
        "required": [
          "proposalId",
          "decision"
        ],
        "title": "SchedulePermissionDecision",
        "type": "object"
      },
      "SchedulePermissionRequest": {
        "additionalProperties": false,
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/SchedulePermissionDecision"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Decisions",
            "type": "array"
          }
        },
        "required": [
          "decisions"
        ],
        "title": "SchedulePermissionRequest",
        "type": "object"
      },
      "SchedulePermissionResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SchedulePermissionResult"
            },
            "title": "Results",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SchedulePermissionResponse",
        "type": "object"
      },
      "SchedulePermissionResult": {
        "properties": {
          "applied": {
            "default": false,
            "title": "Applied",
            "type": "boolean"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "proposalId": {
            "title": "Proposalid",
            "type": "string"
          },
          "scope": {
            "title": "Scope"
          },
          "status": {
            "default": "unknown",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "proposalId",
          "decision"
        ],
        "title": "SchedulePermissionResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "面向高校工科学生的 AI 学习与工程实践助教",
    "title": "LingxiLearn",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent-interactions/schedule-permissions": {
      "post": {
        "operationId": "decide_schedule_permissions_api_agent_interactions_schedule_permissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulePermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulePermissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Schedule Permissions"
      }
    },
...
    "/api/agent-tasks/{task_id}/confirmations": {
      "post": {
        "operationId": "confirm_agent_work_api_agent_tasks__task_id__confirmations_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConfirmation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmWorkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Agent Work"
      }
    },
...
    "/api/agent-tasks/{task_id}/interactions/{interaction_id}/answers": {
      "post": {
        "operationId": "answer_agent_interaction_api_agent_tasks__task_id__interactions__interaction_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interaction_id",
            "required": true,
            "schema": {
              "title": "Interaction Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInteractionAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionAnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Agent Interaction"
      }
    },
...
    "/api/agent-tasks/{task_id}/quiz-submissions": {
      "post": {
        "operationId": "submit_agent_quiz_api_agent_tasks__task_id__quiz_submissions_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmissionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Agent Quiz"
      }
    },
...
  }
}